Support resizing the captured window
When the window is resized, X11 allocates a new pixmap for it, so we
need to obtain the new handle.

When the window's width shrinks, it messes up the row-major order of the
matrix, so we need to pad each row with zeroes. When the window's height
shrinks, we need to zero out the trailing rows to discard the previous
frame's data. We combine both of these steps. While the height-only case
could be more performant, a window is usually resized in both
directions.

When either dimension grows, we don't care. We'll continue to extract
the original size, which will now be only a part of the window.