Don't duplicate resize events, but never ignore them, either
We can drop resize events in the ResizeMonitor itself. If the buffer is
full, the consumer hasn't reacted to the resize event yet, and will see
the latest state once it finally does.

We mustn't, however, drop the event in the capture loop. Dropping damage
events is fine, but dropping a resize event in favour of a damage event
means we'll not always obtain a new pixmap.